Add unit tests and refactor code
Made the YelpRequestController and YelpResult Autowired to follow
dependency injection pattern

Switched ExecutorCompletionService from ExecutorService to allow for
dependency injection. Injecting an ExecutorService into the
YelpRequestController causes an error when multiple requests from the
client are made since once ExecutorService is shut down, it cannot be
restarted

Made the Scraper Autowired to follow dependency injection pattern.
Threads no longer create a Scraper for each request, instead they reuse
the same Scraper object

Added a ScrapeResult class that holds the results of a single request
to a Yelp URL. This is the result returned from the scrapeYelp function
in the Scraper class and makes it easier to pass around the data
scraped from a Yelp request

Added unit tests for ApplicationController, YelpRequestController, and
YelpResult