Implement initial and additional requests
Changed so that all requests are either initial or additional requests.
Initial requests are for the pictures in the first page of the photo
gallery. Additional requests are for pictures in the pages after the
first page. This allows for a fast initial request to get the first
30 pictures which can be shown more quickly to the user. An additional
request for more photos can then be made as a background task